+++ BlackbirdDsl/Include/AssemblyInfo.h
@@ -2,9 +2,9 @@
 
 #define Assembly_Copyright L"Copyright © BlackbirdSql 2023"
 #define Assembly_Description L"BlackbirdSql DSL Parser"
-#define File_Version 14,5,3,0
-#define File_Version_Str "14.5.3.0"
+#define File_Version 14,5,3,1
+#define File_Version_Str "14.5.3.1"
 
 //CLR assembly version
-#define Assembly_Version L"14.5.3.0"
+#define Assembly_Version L"14.5.3.1"